server - Copy/requirements.txt
flask==3.1.2
flask-cors==6.0.1
gunicorn==21.2.0
numpy==1.26.4
Pillow==10.3.0
tensorflow-cpu==2.10.0
